+++ NetworkRealtimeWidgets/NetworkRealtimeWidgetsBundle.swift
@@ -1,0 +1,10 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct NetworkRealtimeWidgetsBundle: WidgetBundle {
+  @WidgetBundleBuilder
+  var body: some Widget {
+    SizeWidget()
+  }
+}